requirements.txt
pyyaml
requests
kubernetes
docker-compose
wheel
setuptools
